move UITableViewExtension and TypeAlias to DcCore
