+++ DcCore/DcCore/Helper/TypeAlias.swift
@@ -1,0 +1,3 @@
+import UIKit
+
+public typealias VoidFunction = () -> Void
