--- deltachat-ios/Helper/TypeAlias.swift
@@ -1,3 +1,0 @@
-import UIKit
-
-typealias VoidFunction = () -> Void
